+++ src/odin-api/muninn/client/data.py
@@ -4,7 +4,7 @@ import json
 import asyncio
 import argparse
 import websockets
-from odin.api import ODIN_URL, ODIN_PORT, ODIN_SCHEME, HttpClient, ODIN_API_LOGGER, APIField, APIStatus
+from muninn import ODIN_URL, ODIN_PORT, ODIN_SCHEME, HttpClient, ODIN_API_LOGGER, APIField, APIStatus
 
 
 async def request_data(url: str, resource: str) -> None:
