+++ src/odin-api/muninn/client/ping.py
@@ -5,7 +5,7 @@ import asyncio
 from datetime import datetime
 import json
 import websockets
-from odin.api import ODIN_URL, ODIN_PORT, ODIN_SCHEME, ODIN_API_LOGGER, APIField, APIStatus
+from muninn import ODIN_URL, ODIN_PORT, ODIN_SCHEME, ODIN_API_LOGGER, APIField, APIStatus
 
 
 async def ping(uri: str, message: str) -> None:
