--- src/odin-api/odin/api/version.py
@@ -1,1 +1,0 @@
-__version__ = "0.6.4"
