+++ src/odin-api/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 NAME = odin-api 
-version = 0.6.4
-description = Odin Client API
+version = 0.6.5
+description = Odin Client API (Muninn)
 author = interactions
 author_email = odin@interactions.com
 license = Apache 2.0
@@ -31,20 +31,20 @@ classifiers =
 [options]
 package_dir=
     =.
-packages = odin.api 
+packages = muninn
 scripts =
-    odin/api/client/authenticate.py
-    odin/api/client/cleanup.py
-    odin/api/client/create.py
-    odin/api/client/data.py
-    odin/api/client/events.py
-    odin/api/client/gpus.py
-    odin/api/client/logs.py
-    odin/api/client/ping.py
-    odin/api/client/push.py
-    odin/api/client/run.py
-    odin/api/client/show.py
-    odin/api/client/user.py
+    muninn/client/authenticate.py
+    muninn/client/cleanup.py
+    muninn/client/create.py
+    muninn/client/data.py
+    muninn/client/events.py
+    muninn/client/gpus.py
+    muninn/client/logs.py
+    muninn/client/ping.py
+    muninn/client/push.py
+    muninn/client/run.py
+    muninn/client/show.py
+    muninn/client/user.py
 
 install_requires =
     numpy
@@ -58,19 +58,19 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    odin-auth = odin.api.client.authenticate:main
-    odin-cleanup = odin.api.client.cleanup:main
-    odin-create = odin.api.client.create:main
-    odin-data = odin.api.client.data:main
-    odin-events = odin.api.client.events:main
-    odin-gpus = odin.api.client.gpus:main
-    odin-logs = odin.api.client.logs:main
-    odin-ping = odin.api.client.ping:main
-    odin-push = odin.api.client.push:main
-    odin-run = odin.api.client.run:main
-    odin-show = odin.api.client.show:main
-    odin-status = odin.api.client.status:main
-    odin-user = odin.api.client.user:main
+    odin-auth = muninn.client.authenticate:main
+    odin-cleanup = muninn.client.cleanup:main
+    odin-create = muninn.client.create:main
+    odin-data = muninn.client.data:main
+    odin-events = muninn.client.events:main
+    odin-gpus = muninn.client.gpus:main
+    odin-logs = muninn.client.logs:main
+    odin-ping = muninn.client.ping:main
+    odin-push = muninn.client.push:main
+    odin-run = muninn.client.run:main
+    odin-show = muninn.client.show:main
+    odin-status = muninn.client.status:main
+    odin-user = muninn.client.user:main
 
 [options.extras_require]
 test:
